Trim unintended Environment.NewLine from docker secret contents before adding to Configuration
Docker secrets created on a Linux swarm using `echo "secret contents" | docker secret create test-secret -` will append a newline on the end of the secret contents. In a dotnetcore app running in a dockerized service with this secret exposed to the service, the `DockerSecretsConfigurationProvider` reads the secret with the newline character into the Configuration (resulting in something like "secret contents\n"). This is clearly not the intended secret contents.